django-app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.2
channels>=4.0
channels-redis>=4.1
daphne>=4.0
redis>=4.5
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
django-filter>=23.0
drf-yasg>=1.21

# Development & Testing
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
pytest-asyncio>=0.21
factory-boy>=3.2
coverage>=7.2
black>=23.0
flake8>=6.0